delete lifecycle and fix some bugs

+++ handler/remove_handler.py
@@ -10,18 +10,39 @@ import os
 class RemoveHandler(BaseHandler):
     def post(self):
         """
-        code 101: name not exists
-        code 102: wrong nodes or not exists
-        code 200: ok
+        code 101: blank input
+        code 102: invalid input nodes
+        code 103: name not exists
+        code 200: success
         :return:
         """
-        cname = self.get_argument('cname')
-        ret = {}
+        username = self.get_argument('cname')
+        str_nodes = self.get_argument('nodes')
 
-        for node_id in range(1, 18 + 1):
-            container_name = '%s.node%.2d' % (cname, node_id)
+        ret = {'code': None}
+        if username == '' or str_nodes == '':
+            ret['code'] = 101
+            self.write(ret)
+            return
+
+        node_list = self.get_node_list_by_str_nodes(str_nodes)
+        if node_list == None:
+            ret['code'] = 102
+            self.write(ret)
+            return
+
+        uid = self.db.get_uid_by_username(username)
+        if not uid:
+            ret['code'] = 103
+            self.write(ret)
+            return
+
+        for node_id in node_list:
+            container_name = '%s.node%.2d' % (username, node_id)
             os.system('ssh node%.2d "docker stop %s && docker rm %s"' % (node_id, container_name, container_name))
 
             print('close', container_name, 'done')
-        self.redirect('/')
 
+        self.db.remove_user_permission(uid, node_list)
+        ret['code'] = 200
+        self.write(ret)
